Create README.md
Employee attrition refers to the rate at which employees leave an organization over a certain period of time. It is a crucial metric that measures the number or percentage of employees who voluntarily or involuntarily separate from organization within a specific timeframe. Addressing this issue is essential for the growth of an organization.

This project aims to explain the analysis of employee dataset from the hiring department of an organization, providing valuable insights and recommendations regarding employee attrition. By identifying employees who are at risk of attrition and implementing effective retention strategies.